@@ -6,25 +6,23 @@ authors = ["Markus Iorio <mbalzert@gmail.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.11"
-fastapi = "^0.97.0"
-sqlalchemy = "^2.0.16"
-pydantic = { extras = ["dotenv"], version = "^1.10.9" }
-httpx = "^0.24.1"
-uvicorn = "^0.22.0"
+python = "^3.12"
 alembic = "^1.11.1"
-result = "^0.13.1"
+fastapi = "^0.103.1"
+uvicorn = "^0.23.2"
+pydantic-settings = "^2.0.3"
+sqlalchemy = "^2.0.21"
 psycopg2-binary = "^2.9.9"
+httpx = "^0.25.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.0.262"
-black = "^23.3.0"
 mypy = "^1.2.0"
 pre-commit = "^3.2.2"
-blacken-docs = "^1.13.0"
 pytest = "^7.3.2"
 pdoc3 = "^0.10.0"
 coverage = "^7.3.1"
+black = "^23.10.1"
 
 [build-system]
 requires = ["poetry-core"]
@@ -32,7 +30,6 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 119
-target-version = ['py311']
 
 [tool.yapf]
 based_on_style = "pep8"
@@ -87,4 +84,4 @@ select = [
     "UP",
     "YTT",
 ]
-ignore = ["ANN001", "ANN101", "ANN102", "D211", "D212", "D107"]
+ignore = ["ANN001", "ANN101", "ANN102", "D211", "D212", "D107", "Q000"]
