@@ -2,10 +2,10 @@
 import datetime
 import json
 
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, ConfigDict, Field, ValidationInfo, field_validator
 
 
-def _get_now_year() -> int:
+def _get_year_now() -> int:
     return datetime.datetime.now(tz=datetime.UTC).year
 
 
@@ -13,14 +13,17 @@ class VehicleBase(BaseModel):
 
     """Base vehicle model."""
 
-    name: str = Field(description="The name of the vehicle.", example="Audi")
+name: str = Field(description="The name of the vehicle.", examples={'default': 'Audi'})
     year_of_manufacture: int = Field(
         description="The year of manufacture for the vehicle.",
         ge=2000,
-        le=_get_now_year(),
-        default=_get_now_year(),
+        le=_get_year_now(),
+        default=_get_year_now(),
     )
-    ready_to_drive: bool = Field(description="Whether the vehicle is ready to drive.", default=False)
+    ready_to_drive: bool = Field(
+        description="Whether the vehicle is ready to drive.",
+        default=False,
+    )
 
 
 class VehicleCreate(VehicleBase):
@@ -30,12 +33,12 @@ class VehicleCreate(VehicleBase):
     body: dict = Field(
         description="Additional information about the vehicle in the form of a dictionary.",
         default_factory=dict,
-        example=dict(color="black"),
+        examples=[dict(color="black")],
     )
 
-    @validator("body", pre=True)
+    @field_validator("body", mode="before")
     @classmethod
-    def parse_body(cls, v: str | dict) -> dict | None:
+    def parse_body(cls, v: str, _: ValidationInfo) -> dict:
         """
         Parse the body value as JSON.
 
@@ -55,19 +58,26 @@ class VehicleUpdate(BaseModel):
 
     """Vehicle update model."""
 
-    name: str | None = Field(description="The name of the vehicle.", example="Audi", default=None)
+    name: str | None = Field(
+        description="The name of the vehicle.",
+        examples=["Audi"],
+        default=None,
+    )
     year_of_manufacture: int | None = Field(
         description="The year of manufacture for the vehicle.",
-        ge=2000,
-        le=_get_now_year(),
+        ge=1980,
+        le=_get_year_now(),
         default=None,
     )
     body: dict | None = Field(
         description="Additional information about the vehicle in the form of a dictionary.",
         default_factory=dict,
-        example=None,
+        examples=[dict(color="black")],
     )
-    ready_to_drive: bool = Field(description="Whether the vehicle is ready to drive.", default=None)
+    ready_to_drive: bool = Field(
+        description="Whether the vehicle is ready to drive.",
+        default=None,
+    )
 
 
 class VehicleInDBBase(VehicleCreate):
@@ -75,12 +85,7 @@ class VehicleInDBBase(VehicleCreate):
     """Vehicle in DB model."""
 
     id: int | None = None  # noqa: A003
-
-    class Config:
-
-        """Pydantic config."""
-
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
 class Vehicle(VehicleInDBBase):
