@@ -4,9 +4,8 @@
 import datetime
 import logging
 
-from fastapi import APIRouter, Depends, HTTPException, Query, status
+from fastapi import APIRouter, Depends, Query, status
 
-from src.core.error import HTTPError
 from src.core.session import SESSION_LOCAL, AbstractSession
 from src.crud import REPOSITORY_LOCAL, AbstractRepository
 from src.model.vehicle import Vehicle
@@ -26,48 +25,43 @@ def list_vehicle(
     *,
     session: AbstractSession = Depends(SESSION_LOCAL),
     repository: AbstractRepository = Depends(REPOSITORY_LOCAL(Vehicle)),
-    name: str | None = Query(default=None, description=FILTER_ON.format(criterion="vehicle name"), example="Audi"),
+    name: str
+    | None = Query(
+        default=None,
+        description=FILTER_ON.format(criterion="vehicle name"),
+        examples=["Audi"],
+    ),
     year_of_manufacture: int
     | None = Query(
         default=None,
         ge=2000,
         le=datetime.datetime.now(tz=datetime.UTC).date().year,
-        example=2020,
+        examples=[2020],
         description=FILTER_ON.format(criterion="year of manufacture"),
     ),
     ready_to_drive: bool
-    | None = Query(default=None, description=FILTER_ON.format(criterion="ready to drive"), example=True),
+    | None = Query(
+        default=None,
+        description=FILTER_ON.format(criterion="ready to drive"),
+        examples=[True],
+    ),
 ) -> list[schemas.Vehicle]:
     """
     List all vehicles.
 
     Filters can be applied to refine results based on name, manufacturing year, and readiness for driving.
     """
-    try:
-        with session as db:
-            vehicles: list[Vehicle] = services.list(
-                db,
-                repository,
-                filter_by=_set_filter(name, year_of_manufacture, ready_to_drive),
-            )
-    except HTTPError as e:
-        raise HTTPException(status_code=e.status_code, detail=e.detail) from e
-    except Exception as e:
-        log.exception(UNCAUGHT)
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR) from e
-    else:
-        return [schemas.Vehicle.from_orm(vehicle) for vehicle in vehicles]
-
-
-def _set_filter(name: str | None, year_of_manufacture: int | None, ready_to_drive: bool | None) -> dict:
-    filter_by = {}
-    if name is not None:
-        filter_by["name"] = name
-    if year_of_manufacture is not None:
-        filter_by["year_of_manufacture"] = year_of_manufacture
-    if ready_to_drive is not None:
-        filter_by["ready_to_drive"] = ready_to_drive
-    return filter_by
+    with session as db:
+        vehicles: list[Vehicle] = services.list(
+            db,
+            repository,
+            filter_by={
+                "name": name,
+                "year_of_manufacture": year_of_manufacture,
+                "ready_to_drive": ready_to_drive,
+            },
+        )
+    return [schemas.Vehicle.model_validate(vehicle) for vehicle in vehicles]
 
 
 @router.post("/", response_model=schemas.Vehicle)
@@ -89,16 +83,12 @@ def create_vehicle(
     ready_to_drive: A boolean flag indicating whether the vehicle is ready to drive.
     Defaults to False.
     """
-    try:
-        with session as db:
-            vehicle: Vehicle = services.create(db, repository, to_create=to_create)
-    except HTTPError as e:
-        raise HTTPException(status_code=e.status_code, detail=e.detail) from e
-    except Exception as e:
-        log.exception(UNCAUGHT)
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR) from e
-    else:
-        return schemas.Vehicle.from_orm(vehicle)
+    with session as db:
+        return schemas.Vehicle.model_validate(services.create(
+            db,
+            repository,
+            to_create=to_create,
+        ))
 
 
 @router.put("/{id}", response_model=schemas.Vehicle)
@@ -117,16 +107,8 @@ def update_vehicle(
     id: The ID of the vehicle to update.\
     update_with: An instance of `schemas.VehicleUpdate` with updated information.
     """
-    try:
-        with session as db:
-            vehicle: Vehicle = services.update(db, repository, id, update_with)
-    except HTTPError as e:
-        raise HTTPException(status_code=e.status_code, detail=e.detail) from e
-    except Exception as e:
-        log.exception(UNCAUGHT)
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR) from e
-    else:
-        return schemas.Vehicle.from_orm(vehicle)
+    with session as db:
+        return schemas.Vehicle.model_validate(services.update(db, repository, id, update_with))
 
 
 @router.get("/{id}", response_model=schemas.Vehicle)
@@ -143,16 +125,8 @@ def get_vehicle(
     ----
     id: The ID of the vehicle to retrieve.
     """
-    try:
-        with session as db:
-            vehicle: Vehicle = services.get(db, repository, id, Vehicle())
-    except HTTPError as e:
-        raise HTTPException(status_code=e.status_code, detail=e.detail) from e
-    except Exception as e:
-        log.exception(UNCAUGHT)
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR) from e
-    else:
-        return schemas.Vehicle.from_orm(vehicle)
+    with session as db:
+        return schemas.Vehicle.model_validate(services.get(db, repository, id, Vehicle()))
 
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
@@ -169,11 +143,5 @@ def delete_vehicle(
     ----
     id: The ID of the vehicle to delete.
     """
-    try:
-        with session as db:
-            services.delete(db, repository, id)
-    except HTTPError as e:
-        raise HTTPException(status_code=e.status_code, detail=e.detail) from e
-    except Exception as e:
-        log.exception(UNCAUGHT)
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR) from e
+    with session as db:
+        services.delete(db, repository, id)
