@@ -4,13 +4,22 @@ from typing import TypeVar
 
 from src.core.error import HTTPError
 from src.core.session import AbstractSession
-from src.crud import AbstractRepository, CreateSchemaType, ModelType, UpdateSchemaType
+from src.crud import (
+    AbstractRepository,
+    CreateSchemaType,
+    ModelType,
+    UpdateSchemaType,
+)
 
 UNPROCESSABLE = "unprocessable value, not a"
 T = TypeVar("T")
 
 
-def create(session: AbstractSession, repository: AbstractRepository, to_create: CreateSchemaType) -> ModelType:
+def create(
+    session: AbstractSession,
+    repository: AbstractRepository,
+    to_create: CreateSchemaType,
+) -> ModelType:
     """
     Create a new vehicle.
 
@@ -31,7 +40,12 @@ def create(session: AbstractSession, repository: AbstractRepository, to_create: 
     return repository.create(session=session, to_create=to_create)
 
 
-def get(session: AbstractSession, repository: AbstractRepository, id: int, default: T | None = None) -> ModelType:
+def get(
+    session: AbstractSession,
+    repository: AbstractRepository,
+    id: int,
+    default: T | None = None,
+) -> ModelType:
     """
     Get a vehicle.
 
@@ -75,7 +89,11 @@ def list(  # noqa: A001
     -------
     returns: A list of `Vehicle` objects.
     """
-    return repository.list(session, filter_by=filter_by)
+    return repository.list(session, filter_by=_remove_none_values(filter_by))
+
+
+def _remove_none_values(dictionary: dict) -> dict:
+    return {k: v for k, v in dictionary.items() if v is not None}
 
 
 def update(
